fix(spoke-operations): address PR review parity and OCC safety
Compare revisions as u64, reject revision overflow on promote, align scope
empty-array semantics with TypeScript, add wire validators for computable ops,
and reject invalid assemble extension namespace keys instead of dropping them.

+++ crates/spoke-operations/src/upsert.rs
@@ -151,10 +151,7 @@ fn validate_update_path(candidate: &KnowledgeEntry, stored: &KnowledgeEntry) -> 
         );
     };
 
-    assert_revision_match(
-        revision as f64,
-        stored.revision.unwrap_or(0) as f64,
-    )
+    assert_revision_match(revision, stored.revision.unwrap_or(0))
 }
 
 /// Validate KnowledgeEntry upsert before persist; create vs update inferred from stored presence.
